third_party/WebKit/LayoutTests/platform/win/external/csswg-test/css-flexbox-1/getcomputedstyle/flexbox_computedstyle_flex-basis-0percent-expected.txt

This is a testharness.js-based test.
FAIL flexbox | computed style | flex-basis: 0% assert_equals: expected "0px" but got "0%"
Harness: the test ran to completion.
